fix: use commandline flag for world dir
due to a regression in spigot server [spigot jira
issue](https://hub.spigotmc.org/jira/browse/SPIGOT-5824) the config file
does not work for specifying the world container directory meaning that
everything is saved into the current working directory of the minecraft
jars. This meant that all of the backups are useless.